Move `process_command_result` to separate file
