+++ crates/sncast/src/helpers/mod.rs
@@ -1,6 +1,7 @@
 pub mod account;
 pub mod block_explorer;
 pub mod braavos;
+pub mod command_processing;
 pub mod config;
 pub mod configuration;
 pub mod constants;
